Handle hitting enter with empty, nested list items
If we "deeply" nest a list item (e.g. repeatedly indent it), then
hit enter, it should be deleted as usual. However, the actual behavior
was that the cursor would move to a new line and the previous line's
indent would be decreased.

This is because the list item was deleted, but it was only one item in a
chain of list/listitem/list/listitem for each level of indentation. We
can solve this by traversing the tree to find the root of the indented
subtree, and targeting that node.