Maintain indentation when creating newlines in codeblocks (#1124)
As an implementation detail, this method allows insertNewAfter on
ElementNode to return any LexicalNode. This allows greater control
over what's inserted into the document and the final selection state
after insertNewAfter; previously, if an ElementNode was not returned,
a newline was added and a default selection made.

This change is breaking for this reason (as callers may have made
assumptions based on the previous return type).