Handle pressing enter on an empty list bullet (#1310)
* Handle pressing enter on an empty list bullet

This was tricker than I initially expected, because I rather than the
logic running inside the ListItemNode's `insertNewAfter`, we need to run
it as a command handler for `insertParagraph`. There are cases when the
user makes the gesture "inesrt paragraph" and nothing is inserted, only
removed and moved around.

The actual behavior, expressed as clearly as possible:
If a user hits enter with their cursor in an empty list item, that list
item will lose a level of indentation. If that would be its last level
of indentation, the list item is deleted and replaced by a paragraph.

Any siblings of the list item that come after it will need to be moved
to a new location, either within the existing list or in a new list.

If the list item's parent is now empty, it should be deleted. The same
for that list item's parent and so on, to avoid leaving behind artifacts
of empty lists.

This new logic also obsoletes some old logic, which handled specifically
the case of the *last* list item being deleted.

* Remove now-outdated unit tests

* Remove outdated test

* Use skipIf for the jest test