Fix insertNodes when replacing content adjacent to an inline element
Fixes #3136.

This patch adds six unit tests.

The two "with a fully-selected text node" tests are regression tests and
pass with and without the code changes in this patch.

The other unit tests fail without the code changes in this patch.

The added E2E tests produce a similar scenario to the unit tests and
fail without the code changes from this page.