Updated link markdown regex
no issue

The current `regExp` is causing non-link Markdown syntax to be parsed as a link.
Example, the markdown syntax for an image would be `![alt text](https://example.com/example.jpeg`
would be detected as link markdown.
This fix updates the regex to not render a link if ! is infront of the
link markdown syntax.