MarkdownExport: fix export for overlapping styles
Previously, exporting text with "layered" formats resulted in layered
markdown. For example, this string:

     root
      └ (53) paragraph
        ├ (48) text  "he"
        ├ (49) text  "llo" { format: bold }
        ├ (50) text  "wor" { format: bold, italic }
        ├ (51) text  "ld" { format: italic }
    >   └ (52) text  "!"

Would be exported like this:

    he**llo*wor**ld*!

The export incrementally applied the italic to a portion of a bold span.
This seems to not be valid according to Commonmark, although micromark
supports it to some extent (but not fully -- it breaks on strikethroughs
for me).

This change closes and opens markdown tags every time the format
changes between text nodes. It now exports that example like this:

    he**llo*****wor****ld*!

While this uses a few more characters, it will work in more
circumstances, it fixes the specific issue I was having, and it matches
Draft.js's export behavior, anecdotally.

Fixes #4895