Change fork modules to not use dynamic imports with top-level await.
This seems to solve a bunch of compatibility issues I ran into,
the ecosystem just isn't ready for top-level await.

With package.json marked as not having side-effects, it seems this
can be optimized properly (at least by vite, tested w/ SvelteKit).
I verified this by building a SvelteKit app that uses lexical and
searching for strings that would only be in the dev build, e.g.
`should implement "importDOM"`.