Pass click event in read-only editor
Right now only a SELECTION_CHANGE_COMMAND is passed when the editor is in read-only mode. Which makes sense, since we don't want to allow copy, paste, etc when contenteditable doesn't have true 'readonly' mode. 

However, I stumbled upon a situation where this logic disabled the 'click' event not firing the CLICK_COMMAND at all when in read-only, which is needed if you want to have any interactivity even in read-only when a non-editing user needs to interact with the widget, eg open a popup on a double click to view something for example. So passing this event through irrespective of editor editability.