Update index.tsx
When using the playground in a smaller window, and scrolling down, it is impossible to move a node to the top node.

This fix adjusts the draggableBlockPlugin to allow the placement of a node at the root while scrolled. I encountered this bug in my own build of this plugin.