Correctly take zoom into account when calculating table drag zones
In our implementation of this editor, the user can modify the zoom of
the editor area independently of the rest of the page. This skews the
selection areas of the table cell resizer. This commit fixes that.

It also slightly shifts the rectangles to *center* on the edge,
instead of being outside the cell, and scales these rectangles
proportionally. This makes for a more intuitive user experience as the
user will try to select the edge, not the outside.