Fix `RangeSelection#extract` Bug
When `$isTextNode(firstNode)` is true above, it's possible to run `selectedNodes.shift()`. Then `lastIndex` is not longer the last index, but rather the index after the end of the array - resulting in this line appending to the current array rather than replacing as is intended.

Use the current length of `selectedNodes` to avoid this edge case.