Always Update `_cachedNodes` on `RangeSelection#extract`
`selectedNodes` is a reference to `_cachedNodes`, so changes are reflected. Except, during `splitText`, `_cachedNodes` is reset to null, so later changes will not be reflected. (The reset is due to assigning to the `anchor`/`focus` of the current selection - `Point#set` resets `_cachedNodes` on the associated selection).

It's very confusing that sometimes the selection will have its nodes updated on `extract` and not others. In fact, in this edge case, while the first run of `extract` will not update `_cachedNodes`, an immediate subsequent run will.