Ensure First Match is Used in `importTextMatchTransformer`
Ensure we use the first match if there are multiple on a line (when "outermost" is ambiguous), rather than just the match from the first transformer that ran on the line.

This prevents cases where there is a `TextFormatTransformer` match in an earlier `TextMatchTransformer`. The `TextFormatTransformer` will segment the earlier `TextMatchTransformer` so that it is never parsed correct. See added test for an example that failed before this fix (the link would not be parsed and the raw link markdown was present in the final output).