TypeaheadMenuPlugin: Add an option to not reposition the menu.
By default, the TypeaheadMenuPlugin positions the menu below the trigger character. Sometimes it is useful to chose a different, possibly fixed positioning e.g. above or below the input. LexicalMenu supports this by allowing to pass null as `resolution` to `useMenuAnchorRef`. This change adds an option to make this accessible when using TypeaheadMenuPlugin.
The default behavior does not change.